20-de-nov

+++ 01_ejemplos/area.cpp
@@ -3,22 +3,22 @@
 #include <stdlib.h>
 
 int main(){
-    int lado;
-    int area, perimetro;
+    double lado;
+    double area, perimetro;
 
     system("toilet --metal -fpagga CUADRADO");
 
     /* Entrada de datos */
     printf("Lado: ");
-    scanf(" %i", &lado);
+    scanf(" %lf", &lado);
 
     /* Calculos */
     area = lado * lado;
     perimetro = 4 * lado;
 
     /* Salida de datos */
-    printf("area: %i\n"
-           "perimetro: %i\n",
+    printf("area: %lf\n"
+           "perimetro: %lf\n",
            area, perimetro);
     return EXIT_SUCCESS;
